split b2b folder in to smoke and regression
