--- projects/storefrontapp-e2e-cypress/cypress/integration/b2b/my-company/cost-centers.e2e-spec.ts
@@ -1,4 +1,0 @@
-import { costCenterConfig } from '../../../helpers/b2b/my-company/config/cost-center.config';
-import { testMyCompanyFeatureFromConfig } from '../../../helpers/b2b/my-company/my-company.utils';
-
-testMyCompanyFeatureFromConfig(costCenterConfig);
