--- projects/storefrontapp-e2e-cypress/cypress/integration/b2b/my-company/units.e2e-spec.ts
@@ -1,4 +1,0 @@
-import { unitConfig } from '../../../helpers/b2b/my-company/config/unit';
-import { testMyCompanyFeatureFromConfig } from '../../../helpers/b2b/my-company/my-company.utils';
-
-testMyCompanyFeatureFromConfig(unitConfig);
