+++ projects/storefrontapp-e2e-cypress/cypress/integration/b2b/regression/my-company/budgets.e2e-spec.ts
@@ -1,0 +1,4 @@
+import { budgetConfig } from '../../../../helpers/b2b/my-company/config/budget.config';
+import { testMyCompanyFeatureFromConfig } from '../../../../helpers/b2b/my-company/my-company.utils';
+
+testMyCompanyFeatureFromConfig(budgetConfig);
