+++ projects/storefrontapp-e2e-cypress/cypress/integration/b2b/regression/my-company/purchase-limits.e2e-spec.ts
@@ -1,0 +1,6 @@
+import { purchaseLimitConfigs } from '../../../../helpers/b2b/my-company/config/purchase-limit';
+import { testMyCompanyFeatureFromConfig } from '../../../../helpers/b2b/my-company/my-company.utils';
+
+purchaseLimitConfigs.forEach((config) => {
+  testMyCompanyFeatureFromConfig(config);
+});
