+++ projects/storefrontapp-e2e-cypress/cypress/integration/b2b/regression/my-company/units.e2e-spec.ts
@@ -1,0 +1,4 @@
+import { unitConfig } from '../../../../helpers/b2b/my-company/config/unit';
+import { testMyCompanyFeatureFromConfig } from '../../../../helpers/b2b/my-company/my-company.utils';
+
+testMyCompanyFeatureFromConfig(unitConfig);
