+++ projects/storefrontapp-e2e-cypress/cypress/integration/b2b/regression/my-company/user-groups.e2e-spec.ts
@@ -1,0 +1,4 @@
+import { userGroupConfig } from '../../../../helpers/b2b/my-company/config/user-group';
+import { testMyCompanyFeatureFromConfig } from '../../../../helpers/b2b/my-company/my-company.utils';
+
+testMyCompanyFeatureFromConfig(userGroupConfig);
