+++ projects/storefrontapp-e2e-cypress/cypress/integration/b2b/regression/my-company/users.e2e-spec.ts
@@ -1,0 +1,4 @@
+import { userConfig } from '../../../../helpers/b2b/my-company/config/user';
+import { testMyCompanyFeatureFromConfig } from '../../../../helpers/b2b/my-company/my-company.utils';
+
+testMyCompanyFeatureFromConfig(userConfig);
