+++ projects/storefrontapp-e2e-cypress/cypress/integration/b2b/regression/replenishment/b2b-replenishment-order-details-flow.e2e-spec.ts
@@ -1,10 +1,10 @@
-import { loginB2bUser } from '../../../helpers/b2b/b2b-checkout';
-import * as replenishmentDetails from '../../../helpers/b2b/b2b-replenishment-order-details';
+import { loginB2bUser } from '../../../../helpers/b2b/b2b-checkout';
+import * as replenishmentDetails from '../../../../helpers/b2b/b2b-replenishment-order-details';
 import {
   ORDER_REQUEST_ENDPOINT,
   POWERTOOLS_BASESITE,
   USER_REQUEST_ENDPOINT,
-} from '../../../sample-data/b2b-checkout';
+} from '../../../../sample-data/b2b-checkout';
 
 describe('Replenishment order details', () => {
   before(() => {
