+++ projects/storefrontapp-e2e-cypress/cypress/integration/b2b/regression/replenishment/b2b-replenishment-order-history-no-orders-flow.e2e-spec.ts
@@ -1,11 +1,11 @@
-import { replenishmentOrderHistoryUrl } from '../../../helpers/b2b/b2b-replenishment-order-history';
-import { orderHistoryTest } from '../../../helpers/order-history';
+import { replenishmentOrderHistoryUrl } from '../../../../helpers/b2b/b2b-replenishment-order-history';
+import { orderHistoryTest } from '../../../../helpers/order-history';
 import {
   b2bAccountShipToUser,
   ORDER_REQUEST_ENDPOINT,
   POWERTOOLS_BASESITE,
   USER_REQUEST_ENDPOINT,
-} from '../../../sample-data/b2b-checkout';
+} from '../../../../sample-data/b2b-checkout';
 
 describe('Order History with no orders', () => {
   before(() => {
