+++ projects/storefrontapp-e2e-cypress/cypress/integration/b2b/smoke/b2b-credit-card-checkout-flow.e2e-spec.ts
